refactor: main FeedCard 리팩토링
